Played with group associated to directories
After an interesting conversation with Nathan, I wanted to try how
convenient it would be to tie groups to specific folder in the file
system as a way to organize source files.
The idea is to leverage the "Relative to" and see if in mid-long term
it's a viable option for bigger projects.